[Feat]: 검색 결과 페이지- ai를 이용해서 추천 이유 기능 추가
