+++ app/search/result/page.tsx
@@ -8,18 +8,19 @@ import SearchResultProductList from '@/components/search/SearchResultProductList
 import { getProductDetail } from '@/lib/api/products';
 import { useSearchStore } from '@/store/searchStore';
 import { ProductList } from '@/types/product';
+import Image from 'next/image';
 import { useSearchParams } from 'next/navigation';
 import { useEffect, useState } from 'react';
 
 export default function SearchResultPage() {
-  const [proudcts, setProducts] = useState<ProductList[]>([]);
+  const [products, setProducts] = useState<ProductList[]>([]);
   const [isLoading, setIsLoading] = useState(true);
 
-  // url의 id 정보가 담긴 쿼리 파리미터 가져오기
   const searchParams = useSearchParams();
 
-  // 전역으로 저장한 useSearchStore의 query 값을 가져옴
+  // 전역 상태에서 가져오기
   const queryText = useSearchStore(s => s.query);
+  const recommendationReason = useSearchStore(s => s.recommendationReason);
 
   // 상품 정보 가져오기
   useEffect(() => {
@@ -28,10 +29,7 @@ export default function SearchResultPage() {
 
     const loadProducts = async () => {
       try {
-        // result: Promise 5개가 배열로 있음(아직 상품 정보 아님 -> 아직 pending 상태!)
-        // await Promise.all()로 모든 Promise 완료 대기 후 실제 상품 정보 추출
         const result = idArray.map(async id => {
-          // 상품 상세 정보 불러오는 api 호출 함수
           const detailRes = await getProductDetail(id);
 
           if (!detailRes.ok) {
@@ -39,10 +37,8 @@ export default function SearchResultPage() {
           }
           return detailRes.item;
         });
-        // productList가 실제 상품 정보를 담고 있는 배열
-        const productList = await Promise.all(result);
-        // null 제거 및 유효한 상품만 필터링
 
+        const productList = await Promise.all(result);
         setProducts(productList);
       } catch (error) {
         console.error('상품 조회 실패', error);
@@ -55,7 +51,6 @@ export default function SearchResultPage() {
     if (idArray.length > 0) {
       loadProducts();
     } else {
-      // ID가 없을 때(검색 결과가 없을 때)도 로딩 종료
       setIsLoading(false);
     }
   }, [searchParams]);
@@ -64,8 +59,7 @@ export default function SearchResultPage() {
     return <Spinner />;
   }
 
-  // 검색 결과가 없을 때
-  if (proudcts.length === 0) {
+  if (products.length === 0) {
     return (
       <div className="font-pretendard">
         <SearchResultHeader queryText={queryText} productsCount={0} />
@@ -85,9 +79,33 @@ export default function SearchResultPage() {
       <div className="font-pretendard pb-15">
         <SearchResultHeader
           queryText={queryText}
-          productsCount={proudcts.length}
+          productsCount={products.length}
         />
-        <SearchResultProductList products={proudcts} />
+
+        {/* AI 추천 이유 박스*/}
+        {recommendationReason && (
+          <div className="font-pretendard mx-4 my-6 py-5 px-6 bg-br-primary-200 rounded-4xl rounded-bl-none ">
+            <div className="flex flex-row gap-3 items-center">
+              <Image
+                src="/icons/aisearch-sparkle.svg"
+                alt="AI 검색"
+                width={25}
+                height={25}
+                aria-hidden="false"
+              />
+              <h3 className="font-medium text-gray-800 text-base">
+                왜 이 상품들을 추천했나요?
+              </h3>
+            </div>
+            <div className="mt-2">
+              <p className="text-gray-700 text-sm leading-relaxed">
+                {recommendationReason}
+              </p>
+            </div>
+          </div>
+        )}
+
+        <SearchResultProductList products={products} />
 
         <UnderBar />
       </div>
