+++ lib/ai/opeinai-client.ts
@@ -1,0 +1,19 @@
+import OpenAI from 'openai';
+
+// OpenAI 클라이언트 설정
+const OPENAI_API_KEY = process.env.OPENAI_API_KEY;
+
+let openaiClient: OpenAI | null = null;
+
+export function getOpenAIClient(): OpenAI {
+  if (!OPENAI_API_KEY) {
+    throw new Error('OPENAI_API_KEY가 설정되지 않았습니다.');
+  }
+
+  // 이미 open api key가 있으면 재사용
+  if (!openaiClient) {
+    openaiClient = new OpenAI({ apiKey: OPENAI_API_KEY });
+  }
+
+  return openaiClient;
+}
