+++ lib/search/similarity.ts
